ao: software volume control
Previous volume control was in AudioThread. Thas not reasonable because
user must use AVPlayer. Now audio samples are scaled in AudioOutput.
Also scale function is updated less.
TODO: if SetVolume feature is set (for example backend supports volume
control or set by user), use backend implemention
