+++ tests/ao/main.cpp
@@ -1,6 +1,6 @@
 /******************************************************************************
     QtAV:  Media play library based on Qt and FFmpeg
-    Copyright (C) 2014 Wang Bin <wbsecg1@gmail.com>
+    Copyright (C) 2014-2015 Wang Bin <wbsecg1@gmail.com>
 
 *   This file is part of QtAV
 
@@ -87,6 +87,7 @@ int main(int argc, char** argv)
             left = (left+1) % kTableSize;
             right = (right+3)% kTableSize;
         }
+        ao->setVolume(2*sin(2.0*M_PI/1000.0*timer.elapsed()));
         ao->play(data);
     }
     ao->close();
